PRISM: dumper; and ability to stop&continue

+++ backend/data_preparation/crawler/prism_crawler.py
@@ -4,14 +4,15 @@ from datetime import datetime, timedelta, timezone
 from ftplib import FTP, error_perm
 from typing import List, Union
 
-import numpy as np
 import rootpath
 
 rootpath.append()
 
 from paths import PRISM_DATA_PATH
+from backend.data_preparation.connection import Connection
 from backend.data_preparation.crawler.crawlerbase import CrawlerBase
 from backend.data_preparation.extractor.bil_extractor import BILExtractor, BILFormat
+from backend.data_preparation.dumper.prism_dumper import PRISMDumper
 
 logging.getLogger('TaskManager')
 
@@ -64,25 +65,51 @@ class PRISMCrawler(CrawlerBase):
             based on RECENT or HISTORICAL
         """
 
+        with Connection() as conn:
+            cur = conn.cursor()
+            cur.execute('select date, ppt, tmax, vpdmax from prism_info')
+            exist_list = cur.fetchall()
+
+            exist_dict = dict()
+            for date, ppt, tmax, vpdmax in exist_list:
+                exist_dict[date] = (ppt, tmax, vpdmax)
+
         date = self.current_date - timedelta(days=1)
         while date >= end_clause:
-            for var in PRISMCrawler.variables:
+
+            print(f'fetch: {date}')
+            # var_dict = dict()
+            for var_idx, var in enumerate(PRISMCrawler.variables):
+                # skip if exist
+                if date in exist_dict and exist_dict[date][var_idx]:
+                    print(f'skip: {date}-{var}')
+                    continue
+
                 saved_filepath = self.crawl(date, var)
                 if saved_filepath:
                     # noinspection PyTypeChecker
                     bil = self.extractor.extract(saved_filepath)  # type: BILFormat
-                    np.save(os.path.splitext(saved_filepath)[0], bil.ndarray)
+                    self.dumper.insert(date, bil.flattened, var)
+                    # var_dict[var] = bil.flattened
+
                     # clean up
                     os.remove(saved_filepath)
 
+            # table = PRISMCrawler.merge_table(date, var_dict)
             # finish crawling a day
             date = date - timedelta(days=1)
 
     def assign_buffer(self, content) -> None:
         self.buffer.append(content)
 
+    @staticmethod
+    def merge_table(date: datetime.date, var_dict: dict):
+        return [(date, gid, var_dict['ppt'][gid], var_dict['tmax'][gid], var_dict['vpdmax'][gid]) for gid in
+                range(len(var_dict['ppt']))]
+
 
 if __name__ == '__main__':
     crawler = PRISMCrawler()
     crawler.set_extractor(BILExtractor())
+    crawler.set_dumper(PRISMDumper())
     crawler.start(datetime.now().date() - timedelta(days=7))
